% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_contrasts.R
\name{avg_contrasts}
\alias{avg_contrasts}
\title{Calculate weighted average contrasts between random forest predictions for target predictor values}
\usage{
avg_contrasts(
  marginal_preds,
  target.vars,
  by = NULL,
  equal.wt = NULL,
  wt = c("iso", "joint", "all"),
  interval = c(0.05, 0.95),
  verbose = TRUE
)
}
\arguments{
\item{marginal_preds}{Object of class \code{treePredictions} generated by \code{tree_predictions}.}

\item{target.vars}{Character string of targeted predictor variable names from which to derive predictions.}

\item{by}{Character string of secondary variables over which the target contrasts should be averaged.}

\item{equal.wt}{Character string of external predictor variable names across which equal weighting will be applied.}

\item{wt}{Character string indicating the type of weighting to be applied to the calculation of mean predictions. See details below.}

\item{interval}{Numeric values specifying the central proportion of predictions to report. Default is 90\%}

\item{verbose}{Logical. Should extra information be printed?}
}
\value{
a \code{data.frame}
}
\description{
Calculate weighted average contrasts between random forest predictions for target predictor values
}
\details{
This is a wrapper function which takes the output of \code{predicted_contrasts} and calculated weighted average contrasts for secondary predictor levels.
}
\examples{
\dontrun{}
}
\author{
Jason Grafmiller
}
