% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_predictions.R
\name{avg_predictions}
\alias{avg_predictions}
\title{Calculate average predictions for target variables in random forest}
\usage{
avg_predictions(
  marginal_preds,
  target.vars,
  equal.wt = NULL,
  wt = c("iso", "joint", "all"),
  interval = c(0.05, 0.95),
  verbose = TRUE
)
}
\arguments{
\item{marginal_preds}{Object of class \code{treePredictions} generated by \code{tree_predictions}.}

\item{target.vars}{Character string of targeted predictor variable names from which to derive predictions.}

\item{equal.wt}{Character string of external predictor variable names across which equal weighting will be applied.}

\item{wt}{Character string indicating the type of weighting to be applied to the calculation of mean predictions. See details below.}

\item{interval}{Numeric values specifying the central proportion of predictions to report. Default is 90\%}

\item{verbose}{Logical. Should extra information be printed?}
}
\value{
A \code{data.frame} including all combinations of values in \code{target.vars}.
\describe{
\item{\code{mean_X_prob}}{The mean of model predictions for level X weighted by the marginal distributions across all predictors other than \code{target.vars} and \code{equal.wt}.}
\item{\code{lower}}{The lower percentile of model predictions set by \code{interval}.}
\item{\code{upper}}{The upper percentile of model predictions set by \code{interval}.}
}
}
\description{
Calculate average predictions for target variables in random forest
}
\details{
Add details here
}
\examples{
\dontrun{
# not run
library(ranger)

rf <- ranger(Species ~ ., iris, probability = T)

tree_preds <- tree_predictions(rf, iris)

avg_predictions(tree_preds, "Sepal.Length")
}
}
\references{
Add references here
}
\author{
Jason Grafmiller
}
