% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_contrasts.R
\name{marginal_contrasts}
\alias{marginal_contrasts}
\title{Calculate contrasts between marginal predictions for predictor values}
\usage{
marginal_contrasts(marginal_preds, var, verbose = FALSE)
}
\arguments{
\item{marginal_preds}{Object of class \code{marginalPreds} generated by \code{marginal_predictions}.}

\item{var}{Character string of the name of the predictor for which to calculate contrasts}

\item{verbose}{Logical indicating whether additional information should be printed}
}
\value{
A \code{data.frame}.
\describe{
}
}
\description{
Calculate contrasts between marginal predictions for predictor values
}
\details{
Add details here
}
\examples{
\dontrun{}
}
\references{
Add references here
}
\author{
Jason Grafmiller
}
