% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_predictions.R
\name{marginal_predictions}
\alias{marginal_predictions}
\title{Get marginal predictions from a random forest}
\usage{
marginal_predictions(
  m,
  data,
  num.trees = 500,
  n.breaks = 10,
  verbose = TRUE,
  breaks = NULL
)
}
\arguments{
\item{m}{Random forest model. Must be of class \code{ranger}.}

\item{data}{Dataset used to fit model}

\item{num.trees}{Number of trees from which to extract predictions. Default is 500.}

\item{n.breaks}{Number of breaks with which to split continuous predictors. Default is 10.}

\item{verbose}{Logical. Should information be printed?}

\item{breaks}{A named list with values representing custom points for which to get predictions for continuous predictors.}

\item{vars}{Character string of targeted predictor variable names from which to derive marginal predictions. Default is all predictors in \code{m} (not recommended).}

\item{ext_vars}{Character string indicating the external variables to take into consideration when weighting}
}
\value{
A \code{list} of class \code{marginalPreds}.
\describe{
\item{\code{predictions}}{A \code{data.table} of the marginal predictions from the model}
\item{\code{model}}{The name of the model}
\item{\code{data}}{The dataset the model was trained on}
\item{\code{variable_names}}{The names of the independent variables in the model}
\item{\code{predicted.outcome}}{The value representing the positive predicted outcome}
\item{\code{n.breaks}}{The number of breakpoints used for binning continuous predictors}
\item{\code{num.trees}}{The number of trees included in \code{predictions}}
}
}
\description{
Get marginal predictions from a random forest
}
\details{
Add details here
}
\examples{
\dontrun{

}
}
\references{
Add references here
}
\author{
Jason Grafmiller
}
