% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicted_contrasts.R
\name{predicted_contrasts}
\alias{predicted_contrasts}
\title{Calculate contrasts between random forest predictions for target predictor values}
\usage{
predicted_contrasts(
  marginal_preds,
  target.vars,
  interval = c(0.05, 0.95),
  verbose = FALSE
)
}
\arguments{
\item{marginal_preds}{Object of class \code{treePredictions} generated by \code{tree_predictions}.}

\item{target.vars}{Character string of the name of the predictor for which to calculate contrasts}

\item{interval}{Numeric values specifying the central proportion of predictions to report. Default is 90\%}

\item{verbose}{Logical indicating whether additional information should be printed}
}
\value{
A \code{data.frame}.
\describe{}
}
\description{
Calculate contrasts between random forest predictions for target predictor values
}
\details{
Add details here
}
\examples{
\dontrun{}
}
\references{
Sönning, Lukas & Jason Grafmiller. 2022. Seeing the wood for the trees: Predictive margins for random forests. Preprint. \emph{PsyArXiv}. \url{https://doi.org/10.31234/osf.io/jr8yk}.
}
\author{
Jason Grafmiller
}
