% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_predictions.R
\name{avg_predictions}
\alias{avg_predictions}
\title{Calculate average predictions for target variables in random forest}
\usage{
avg_predictions(
  marginal_preds,
  target.vars,
  equal.wt = NULL,
  wt = c("iso", "joint", "all", "none"),
  interval = c(0.05, 0.95),
  verbose = TRUE
)
}
\arguments{
\item{marginal_preds}{Object of class \code{treePredictions} generated by \code{tree_predictions}.}

\item{target.vars}{Character string of targeted predictor variable names from which to derive predictions.}

\item{equal.wt}{Character string of external predictor variable names across which equal weighting will be applied.}

\item{wt}{Character string indicating the type of weighting to be applied to the calculation of mean predictions. See details below. Possible values are \code{'iso'} (default), \code{'joint'}, \code{'all'}, \code{'none'}.}

\item{interval}{Numeric values specifying the central proportion of predictions to report. Default is 90\%}

\item{verbose}{Logical. Should extra information be printed?}
}
\value{
A \code{data.frame} including all combinations of values in \code{target.vars}.
\describe{
\item{\code{mean_X_prob}}{The mean of model predictions for level X weighted by the marginal distributions across all predictors other than \code{target.vars} and \code{equal.wt}.}
\item{\code{lower}}{The lower percentile of model predictions set by \code{interval}.}
\item{\code{upper}}{The upper percentile of model predictions set by \code{interval}.}
}
}
\description{
Calculate average predictions for target variables in random forest
}
\details{
Calculates the weighted average predictions for a target predictor in random forest model.

For calculating averages there are four possible weighting schemes (\code{wt}). The default isolated weighting \code{'iso'} uses the independent marginal distributions of the peripheral variables. Essentially this is the expected proportions we'd calculate in a contingency table. For \code{'joint'}, we use the observed distribution of the combinations of peripheral variable values---in effect the observed proportions in a contingency table. For \code{'all'}, we use the observed distributions, but we include the target variables in these counts as well. Finally, \code{'none'} calculates simple unweighted means.
}
\examples{
\dontrun{
# not run
library(ranger)

rf <- ranger(Species ~ ., iris, probability = T)

tree_preds <- tree_predictions(rf, iris)

avg_predictions(tree_preds, "Sepal.Length")
}
}
\references{
Sönning, Lukas & Jason Grafmiller. 2022. Seeing the wood for the trees: Predictive margins for random forests. Preprint. \emph{PsyArXiv}. \url{https://doi.org/10.31234/osf.io/jr8yk}.
}
\author{
Jason Grafmiller
}
