% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_avg.R
\name{marginal_avg}
\alias{marginal_avg}
\title{Calculate average marginal predictions}
\usage{
marginal_avg(
  marginal_preds,
  target_vars,
  ext_vars = NULL,
  wt = c("iso", "joint", "all"),
  interval = c(0.05, 0.95),
  verbose = TRUE
)
}
\arguments{
\item{marginal_preds}{Object of class \code{marginalPreds} generated by \code{marginal_predictions}.}

\item{target_vars}{Character string of targeted predictor variable names from which to derive predictions.}

\item{ext_vars}{Character string of external predictor variable names across which equal weighting will be applied.}

\item{wt}{Character string indicating the type of weighting to be applied to the calculation of mean predictions. See details below.}

\item{interval}{Numeric values specifying the central proportion of predictions to report. Default is 90\%}
}
\value{
A \code{data.frame} including all combinations of values in \code{target_vars}.
\describe{
\item{\code{mean}}{The mean of model predictions weighted by the marginal distributions across all predictors other than \code{target_vars} and \code{ext_vars}.}
\item{\code{lower}}{The lower percentile of model predictions set by \code{interval}.}
\item{\code{upper}}{The upper percentile of model predictions set by \code{interval}.}
}
}
\description{
Calculate average marginal predictions
}
\details{
Add details here
}
\references{
Add references here
}
\author{
Jason Grafmiller
}
